Fix pymol install

+++ bin/pymol-install.sh
@@ -7,13 +7,20 @@ svnpymol=svnpymol
 mkdir -p $prefix
 mkdir -p $HOME/bin
 
-###### Checkout pymol svn
-svn co --quiet svn://svn.code.sf.net/p/pymol/code/trunk/pymol $prefix/$svnpymol
-###### Build and install pymol
-cd $prefix/$svnpymol
-python setup.py build install --home=$prefix --install-lib=$modules --install-scripts=$prefix
+if [ -f "$prefix/pymol" ];
+then
+   echo "pymol already exists. Skipping install"
+else
+   echo "Installing pymol"
 
-export PATH="$prefix:$modules:$PATH"
+   ###### Checkout pymol svn
+   svn co --quiet svn://svn.code.sf.net/p/pymol/code/trunk/pymol $prefix/$svnpymol
+   ###### Build and install pymol
+   cd $prefix/$svnpymol
+   python setup.py build install --home=$prefix --install-lib=$modules --install-scripts=$prefix
+fi
+
 export PYMOL_PATH="$modules"
+sudo ln -s "$prefix/pymol" /usr/bin/pymol
 
 which pymol || echo ok
